Fixed Pinned Events Stuck in Pinned List Problem
Pinned events were becoming stuck in the pinned list if they were removed from the list directly because the list from which they were being deleted from was not being sent back to the root view controller, thus when moving back and forth the deleted events were getting overrides and thus showing up when not wanted. The events are now deleted from both the eventsList in CalendarPinnedListViewController and pinnedEvents in CalendarViewController.